+++ crates/api-desc/src/crypto.rs
@@ -14,9 +14,13 @@
 
 use crate::*;
 
+#[cfg(feature = "api-crypto-ccm")]
 mod ccm;
+#[cfg(feature = "api-crypto-ec")]
 mod ec;
+#[cfg(feature = "api-crypto-gcm")]
 mod gcm;
+#[cfg(feature = "internal-api-crypto-hash")]
 mod hash;
 
 pub(crate) fn new() -> Item {
@@ -24,6 +28,15 @@ pub(crate) fn new() -> Item {
         /// Cryptographic operations.
     };
     let name = "crypto".into();
-    let items = vec![ccm::new(), ec::new(), gcm::new(), hash::new()];
+    let items = vec![
+        #[cfg(feature = "api-crypto-ccm")]
+        ccm::new(),
+        #[cfg(feature = "api-crypto-ec")]
+        ec::new(),
+        #[cfg(feature = "api-crypto-gcm")]
+        gcm::new(),
+        #[cfg(feature = "internal-api-crypto-hash")]
+        hash::new(),
+    ];
     Item::Mod(Mod { docs, name, items })
 }