+++ crates/prelude/src/crypto.rs
@@ -14,7 +14,11 @@
 
 //! Provides API for cryptography.
 
+#[cfg(feature = "api-crypto-ccm")]
 pub mod ccm;
+#[cfg(feature = "api-crypto-ec")]
 pub mod ec;
+#[cfg(feature = "api-crypto-gcm")]
 pub mod gcm;
+#[cfg(feature = "internal-api-crypto-hash")]
 pub mod hash;