+++ crates/scheduler/src/call/crypto/gcm.rs
@@ -12,36 +12,50 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use wasefire_applet_api::crypto::gcm::{self as api, Api, Support};
+use wasefire_applet_api::crypto::gcm::{self as api, Api};
+#[cfg(feature = "board-api-aes256-gcm")]
 use wasefire_board_api::crypto::aead::Api as _;
-use wasefire_board_api::{self as board, Api as Board, Support as _};
+use wasefire_board_api::Api as Board;
+#[cfg(feature = "board-api-aes256-gcm")]
+use wasefire_board_api::{self as board, Support as _};
 
+#[cfg(feature = "board-api-aes256-gcm")]
 use crate::applet::store::MemoryApi;
-use crate::{DispatchSchedulerCall, SchedulerCall, Trap};
+#[cfg(feature = "board-api-aes256-gcm")]
+use crate::Trap;
+use crate::{DispatchSchedulerCall, SchedulerCall};
 
 pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
     match call {
         Api::Support(call) => support(call),
-        Api::TagLength(call) => tag_length(call),
-        Api::Encrypt(call) => encrypt(call),
-        Api::Decrypt(call) => decrypt(call),
+        Api::TagLength(call) => or_trap!("board-api-aes256-gcm", tag_length(call)),
+        Api::Encrypt(call) => or_trap!("board-api-aes256-gcm", encrypt(call)),
+        Api::Decrypt(call) => or_trap!("board-api-aes256-gcm", decrypt(call)),
     }
 }
 
 fn support<B: Board>(call: SchedulerCall<B, api::support::Sig>) {
     let api::support::Params {} = call.read();
-    let support = board::crypto::Aes256Gcm::<B>::SUPPORT;
-    let support = (support.no_copy as u32) << Support::NoCopy as u32
-        | (support.in_place_no_copy as u32) << Support::InPlaceNoCopy as u32;
+    #[cfg(feature = "board-api-aes256-gcm")]
+    let support = {
+        use wasefire_applet_api::crypto::gcm::Support;
+        let support = board::crypto::Aes256Gcm::<B>::SUPPORT;
+        (support.no_copy as u32) << Support::NoCopy as u32
+            | (support.in_place_no_copy as u32) << Support::InPlaceNoCopy as u32
+    };
+    #[cfg(not(feature = "board-api-aes256-gcm"))]
+    let support = 0;
     call.reply(Ok(api::support::Results { support: support.into() }))
 }
 
+#[cfg(feature = "board-api-aes256-gcm")]
 fn tag_length<B: Board>(call: SchedulerCall<B, api::tag_length::Sig>) {
     let api::tag_length::Params {} = call.read();
     let len = (tag_len::<B>() as u32).into();
     call.reply(Ok(api::tag_length::Results { len }))
 }
 
+#[cfg(feature = "board-api-aes256-gcm")]
 fn encrypt<B: Board>(mut call: SchedulerCall<B, api::encrypt::Sig>) {
     let api::encrypt::Params { key, iv, aad, aad_len, length, clear, cipher, tag } = call.read();
     let scheduler = call.scheduler();
@@ -61,6 +75,7 @@ fn encrypt<B: Board>(mut call: SchedulerCall<B, api::encrypt::Sig>) {
     call.reply(results);
 }
 
+#[cfg(feature = "board-api-aes256-gcm")]
 fn decrypt<B: Board>(mut call: SchedulerCall<B, api::decrypt::Sig>) {
     let api::decrypt::Params { key, iv, aad, aad_len, tag, length, cipher, clear } = call.read();
     let scheduler = call.scheduler();
@@ -80,11 +95,13 @@ fn decrypt<B: Board>(mut call: SchedulerCall<B, api::decrypt::Sig>) {
     call.reply(results);
 }
 
+#[cfg(feature = "board-api-aes256-gcm")]
 const fn tag_len<B: Board>() -> usize {
     use typenum::Unsigned;
     <board::crypto::Aes256Gcm<B> as board::crypto::aead::Api<_, _>>::Tag::USIZE
 }
 
+#[cfg(feature = "board-api-aes256-gcm")]
 fn ensure_support<B: Board>() -> Result<(), Trap> {
     match bool::from(board::crypto::Aes256Gcm::<B>::SUPPORT) {
         true => Ok(()),