+++ crates/scheduler/src/call/platform.rs
@@ -12,23 +12,36 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use wasefire_applet_api::platform::{self as api, Api};
+#[cfg(feature = "board-api-platform")]
+use wasefire_applet_api::platform as api;
+use wasefire_applet_api::platform::Api;
+#[cfg(feature = "board-api-platform")]
+use wasefire_board_api as board;
+#[cfg(feature = "board-api-platform")]
 use wasefire_board_api::platform::Api as _;
-use wasefire_board_api::{self as board, Api as Board};
+use wasefire_board_api::Api as Board;
 
+#[cfg(feature = "board-api-platform")]
 use crate::applet::store::MemoryApi;
-use crate::{DispatchSchedulerCall, SchedulerCall};
+use crate::DispatchSchedulerCall;
+#[cfg(feature = "board-api-platform")]
+use crate::SchedulerCall;
 
+#[cfg(feature = "applet-api-platform-update")]
 mod update;
 
 pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
     match call {
+        #[cfg(feature = "applet-api-platform-update")]
         Api::Update(call) => update::process(call),
-        Api::Version(call) => version(call),
-        Api::Reboot(call) => reboot(call),
+        #[cfg(feature = "applet-api-platform")]
+        Api::Version(call) => or_trap!("board-api-platform", version(call)),
+        #[cfg(feature = "applet-api-platform")]
+        Api::Reboot(call) => or_trap!("board-api-platform", reboot(call)),
     }
 }
 
+#[cfg(feature = "board-api-platform")]
 fn version<B: Board>(mut call: SchedulerCall<B, api::version::Sig>) {
     let api::version::Params { ptr, len } = call.read();
     let scheduler = call.scheduler();
@@ -41,6 +54,7 @@ fn version<B: Board>(mut call: SchedulerCall<B, api::version::Sig>) {
     call.reply(results);
 }
 
+#[cfg(feature = "board-api-platform")]
 fn reboot<B: Board>(call: SchedulerCall<B, api::reboot::Sig>) {
     let api::reboot::Params {} = call.read();
     let res = match board::Platform::<B>::reboot() {