+++ crates/scheduler/src/call/radio/ble.rs
@@ -12,23 +12,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use wasefire_applet_api::radio::ble::{self as api, Advertisement, Api};
+use wasefire_applet_api::radio::ble::Api;
+#[cfg(feature = "board-api-radio-ble")]
+use wasefire_applet_api::radio::ble::{self as api, Advertisement};
+#[cfg(feature = "board-api-radio-ble")]
 use wasefire_board_api::radio::ble::{Api as _, Event};
-use wasefire_board_api::{self as board, Api as Board};
+use wasefire_board_api::Api as Board;
+#[cfg(feature = "board-api-radio-ble")]
+use wasefire_board_api::{self as board};
 
+#[cfg(feature = "board-api-radio-ble")]
 use crate::applet::store::MemoryApi;
-use crate::event::radio::ble::Key;
-use crate::event::Handler;
-use crate::{DispatchSchedulerCall, SchedulerCall, Trap};
+#[cfg(feature = "board-api-radio-ble")]
+use crate::event::{radio::ble::Key, Handler};
+use crate::DispatchSchedulerCall;
+#[cfg(feature = "board-api-radio-ble")]
+use crate::{SchedulerCall, Trap};
 
 pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
     match call {
-        Api::Register(call) => register(call),
-        Api::Unregister(call) => unregister(call),
-        Api::ReadAdvertisement(call) => read_advertisement(call),
+        // TODO: Instead of trapping, we should provide a way to know if supported.
+        Api::Register(call) => or_trap!("board-api-radio-ble", register(call)),
+        Api::Unregister(call) => or_trap!("board-api-radio-ble", unregister(call)),
+        Api::ReadAdvertisement(call) => or_trap!("board-api-radio-ble", read_advertisement(call)),
     }
 }
 
+#[cfg(feature = "board-api-radio-ble")]
 fn register<B: Board>(mut call: SchedulerCall<B, api::register::Sig>) {
     let api::register::Params { event, handler_func, handler_data } = call.read();
     let inst = call.inst();
@@ -47,6 +57,7 @@ fn register<B: Board>(mut call: SchedulerCall<B, api::register::Sig>) {
     call.reply(results);
 }
 
+#[cfg(feature = "board-api-radio-ble")]
 fn unregister<B: Board>(mut call: SchedulerCall<B, api::unregister::Sig>) {
     let api::unregister::Params { event } = call.read();
     let scheduler = call.scheduler();
@@ -59,6 +70,7 @@ fn unregister<B: Board>(mut call: SchedulerCall<B, api::unregister::Sig>) {
     call.reply(results);
 }
 
+#[cfg(feature = "board-api-radio-ble")]
 fn read_advertisement<B: Board>(mut call: SchedulerCall<B, api::read_advertisement::Sig>) {
     let api::read_advertisement::Params { ptr } = call.read();
     let scheduler = call.scheduler();
@@ -71,6 +83,7 @@ fn read_advertisement<B: Board>(mut call: SchedulerCall<B, api::read_advertiseme
     call.reply(results);
 }
 
+#[cfg(feature = "board-api-radio-ble")]
 fn convert_event(event: u32) -> Result<Event, Trap> {
     Ok(match api::Event::try_from(event)? {
         api::Event::Advertisement => Event::Advertisement,