+++ crates/scheduler/src/call/store/fragment.rs
@@ -12,24 +12,30 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use core::ops::Range;
-
-use wasefire_applet_api::store::fragment::{self as api, Api};
+use wasefire_applet_api::store::fragment::Api;
+#[cfg(feature = "board-api-storage")]
+use wasefire_applet_api::store::fragment::{self as api};
 use wasefire_board_api::Api as Board;
+#[cfg(feature = "board-api-storage")]
 use wasefire_store::fragment;
 
+#[cfg(feature = "board-api-storage")]
 use super::convert;
+#[cfg(feature = "board-api-storage")]
 use crate::applet::store::MemoryApi;
-use crate::{DispatchSchedulerCall, SchedulerCall, Trap};
+use crate::DispatchSchedulerCall;
+#[cfg(feature = "board-api-storage")]
+use crate::SchedulerCall;
 
 pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
     match call {
-        Api::Insert(call) => insert(call),
-        Api::Remove(call) => remove(call),
-        Api::Find(call) => find(call),
+        Api::Insert(call) => or_trap!("board-api-storage", insert(call)),
+        Api::Remove(call) => or_trap!("board-api-storage", remove(call)),
+        Api::Find(call) => or_trap!("board-api-storage", find(call)),
     }
 }
 
+#[cfg(feature = "board-api-storage")]
 fn insert<B: Board>(mut call: SchedulerCall<B, api::insert::Sig>) {
     let api::insert::Params { keys, ptr, len } = call.read();
     let scheduler = call.scheduler();
@@ -46,6 +52,7 @@ fn insert<B: Board>(mut call: SchedulerCall<B, api::insert::Sig>) {
     call.reply(results);
 }
 
+#[cfg(feature = "board-api-storage")]
 fn remove<B: Board>(mut call: SchedulerCall<B, api::remove::Sig>) {
     let api::remove::Params { keys } = call.read();
     let results = try {
@@ -58,6 +65,7 @@ fn remove<B: Board>(mut call: SchedulerCall<B, api::remove::Sig>) {
     call.reply(results);
 }
 
+#[cfg(feature = "board-api-storage")]
 fn find<B: Board>(mut call: SchedulerCall<B, api::find::Sig>) {
     let api::find::Params { keys, ptr: ptr_ptr, len: len_ptr } = call.read();
     let scheduler = call.scheduler();
@@ -81,10 +89,11 @@ fn find<B: Board>(mut call: SchedulerCall<B, api::find::Sig>) {
     call.reply(results);
 }
 
-fn decode_keys(keys: u32) -> Result<Range<usize>, Trap> {
+#[cfg(feature = "board-api-storage")]
+fn decode_keys(keys: u32) -> Result<core::ops::Range<usize>, crate::Trap> {
     if keys & 0xf000f000 == 0 {
         Ok((keys & 0xffff) as usize .. ((keys >> 16) & 0xffff) as usize)
     } else {
-        Err(Trap)
+        Err(crate::Trap)
     }
 }